+++ progress/angular/components/src/lib/battery/battery.component.ts
@@ -42,7 +42,12 @@ import { rangeValue } from '../utilities';
                         <path overflow="visible" [attr.d]="getClipPath()" />
                     </clipPath>
                 </defs>
-                <path *ngIf="backgroundColor" [attr.d]="basePath" [attr.fill]="backgroundColor" />
+                <path
+                    *ngIf="backgroundColor"
+                    [attr.d]="basePath"
+                    [attr.fill]="backgroundColor"
+                    [attr.clip-path]="outlined ? undefined : 'url(#' + getID() + ')'"
+                />
                 <path
                     [attr.fill]="(!outlined && backgroundColor) || color || 'currentColor'"
                     [attr.fill-opacity]="outlined || percent >= 100 || (!outlined && backgroundColor) ? '1' : '0.3'"