+++ progress/react/icons/src/types.ts
@@ -1,19 +1,20 @@
 import { SVGAttributes } from 'react';
 
 export type ProgressIconProps = SVGAttributes<SVGElement> & {
-    percent?: number;
-    size?: number;
-    outlined?: boolean;
-    color?: string;
-    showPercentLabel?: boolean;
-    labelPosition?: 'top' | 'bottom' | 'left' | 'right';
-    labelColor?: string;
-    labelSize?: number;
+    backgroundColor?: string;
     classes?: {
         progressIcon?: string;
         progressIconLabel?: string;
         centered?: string;
     };
+    color?: string;
+    labelColor?: string;
+    labelPosition?: 'top' | 'bottom' | 'left' | 'right';
+    labelSize?: number;
+    outlined?: boolean;
+    percent?: number;
+    showPercentLabel?: boolean;
+    size?: number;
 };
 
 export type PieProgressProps = ProgressIconProps & {